% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustModel_func.R
\name{AdjustModel_func}
\alias{AdjustModel_func}
\title{Adjust the model function}
\usage{
AdjustModel_func(model)
}
\arguments{
\item{model}{The original string of users' defined.}
}
\value{
dataframe data.frame(model)
}
\description{
Adjust the model function
}
