% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDe_func.R
\name{CalcDe_func}
\alias{CalcDe_func}
\title{Choose the core code by the model.}
\usage{
CalcDe_func(
  data,
  model,
  guess = c(0, 0, 0, 0, 0, 0),
  method,
  chooseModel = NULL,
  guess2
)
}
\arguments{
\item{data}{Users' data}

\item{model}{The original string users defined.}

\item{guess}{The guess values users input.}

\item{method}{The method users selected.}

\item{chooseModel}{c(2,1): Bivariate first-order differential equation; c(1,2): Univariable second-order differential equation.}

\item{guess2}{Use for Multilevel}
}
\value{
The result of optimization,SE,RMSE,r-squared,users' data,predictor data and output table.
}
\description{
Choose the core code by the model.
}
