% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\title{Univariate second-order differential equation}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{seq}{sequence of observations}
  \item{myTime}{timestamp of observations; the first value of the time variable must be the initial time.}
  \item{myX}{the observed scores}
}
}
\usage{
example1
}
\description{
A dataset containing the myX and time of almost 100 example.
 The variables are as follows:
}
\keyword{datasets}
