% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example3}
\alias{example3}
\title{University of Michigan consumer sentiment index}
\format{
A data frame with 540 rows and 6 variables:
\describe{
  \item{seq}{sequence of observations}
  \item{month}{months of data}
  \item{year}{from 1978 to 2022}
  \item{current}{the Index of Current Economic Conditions (ICC)}
  \item{expected}{the Index of Consumer Expectations (ICE)}
  \item{myTime}{months converted to time series}
}
}
\source{
{ University of Michigan, Survey Research Center, Surveys of Consumers. https://data.sca.isr.umich.edu/}
}
\usage{
example3
}
\description{
The Surveys of Consumers are conducted by the Survey Research Center at the University of
Michigan. Founded in 1946 by George Katona, the surveys have long stressed the important
influence of consumer spending and saving decisions in determining the course of the
national economy.
}
\keyword{datasets}
