% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info_func.R
\name{Info_func}
\alias{Info_func}
\title{Calculate R-squared RMSE SE and so on.}
\usage{
Info_func(solve_data, userdata, predict, var_model, table)
}
\arguments{
\item{solve_data}{a list data, that is a solve of differential equations.}

\item{userdata}{a data frame containing all model variables. The "time" column must be included.}

\item{predict}{predict data.}

\item{var_model}{a dataframe containing equations.}

\item{table}{outtable.}
}
\value{
a list
}
\description{
Calculate R-squared RMSE SE and so on.
}
