% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JudgeModel_func.R
\name{JudgeModel_func}
\alias{JudgeModel_func}
\title{Judge the model belongs of N-variable and N-order differential equation.}
\usage{
JudgeModel_func(model)
}
\arguments{
\item{model}{The original string users defined.}
}
\value{
c(2,1): Bivariate first-order DE; c(1,2): Univariable second-order DE.
}
\description{
Judge the model belongs of N-variable and N-order differential equation.
}
