% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_func.R
\name{Plot_func}
\alias{Plot_func}
\title{Draw the diagram of differential equation}
\usage{
Plot_func(
  userdata,
  predict,
  modelDF,
  var_model,
  field_model,
  order_model,
  multi_model
)
}
\arguments{
\item{userdata}{a data frame containing all model variables. The "time" column must be included.}

\item{predict}{predict data.}

\item{modelDF}{a dataframe of full model.}

\item{var_model}{a dataframe containing equations.}

\item{field_model}{the user's data columns.}

\item{order_model}{N-order differential equation.}

\item{multi_model}{TRUE or FALSE}
}
\value{
plot
}
\description{
Draw the diagram of differential equation
}
