% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDe_func.R
\name{PlotDe_func}
\alias{PlotDe_func}
\title{Plot all data}
\usage{
PlotDe_func(userdata, calcdata)
}
\arguments{
\item{userdata}{Users's data.}

\item{calcdata}{Predictor's data.}
}
\value{
The plot of all data.
}
\description{
Plot all data
}
