% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver_BinFirst_func.R
\name{Solver_BinFirst_func}
\alias{Solver_BinFirst_func}
\title{Core code of Binary first-order differential equational}
\usage{
Solver_BinFirst_func(data, model, guess, method)
}
\arguments{
\item{data}{User's data}

\item{model}{Model's class is dataframe.}

\item{guess}{Guess values that contain coefficient and initial values.}

\item{method}{"Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent"}
}
\value{
The result of optimization,SE,RMSE,r-squared,users's data,predictor data and output table.
}
\description{
Core code of Binary first-order differential equational
}
