% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver_MultiBiFirst_func.R
\name{Solver_MultiBiFirst_func}
\alias{Solver_MultiBiFirst_func}
\title{Solver of Multilevel bivariate first-order differential equation}
\usage{
Solver_MultiBiFirst_func(init_list)
}
\arguments{
\item{init_list}{a list of init_func}
}
\value{
a list
}
\description{
Solver of Multilevel bivariate first-order differential equation
}
