% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver_MultiUniSecond_func.R
\name{Solver_MultiUniSecond_func}
\alias{Solver_MultiUniSecond_func}
\title{Solver of Multilevel univariate second-order differential equation}
\usage{
Solver_MultiUniSecond_func(init_list)
}
\arguments{
\item{init_list}{a list of init_func}
}
\value{
a list
}
\description{
Solver of Multilevel univariate second-order differential equation
}
