% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver_UniSecond_func.R
\name{Solver_UniSecond_func}
\alias{Solver_UniSecond_func}
\title{Solver of univariate second-order differential equation}
\usage{
Solver_UniSecond_func(userdata, var_model, guess, method)
}
\arguments{
\item{userdata}{a data frame containing all model variables. The "time" column must be included.}

\item{var_model}{a dataframe containing equations}

\item{guess}{a list or a string. Guess the coefficients or initial values.}

\item{method}{a list or a string. The available options are 'Nelder-Mead','L-BFGS-B','SANN' and 'BFGS'.}
}
\value{
a list
}
\description{
Solver of univariate second-order differential equation
}
