% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2}
\alias{example2}
\title{Bivariate first-order differential equation}
\format{
A data frame with 30 rows and 4 variables:
\describe{
  \item{myTime}{timestamp of observations; the first value of the time variable must be the initial time.}
  \item{myX}{the observed scores of variable X}
  \item{myY}{the observed scores of variable Y}
}
}
\usage{
example2
}
\description{
A dataset containing the myX, myY and time of almost 30 example.
 The variables are as follows:
}
\keyword{datasets}
