% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDe_func.R
\name{PlotDe_func}
\alias{PlotDe_func}
\title{Plot all data}
\usage{
PlotDe_func(userdata, predictor, model, chooseModel)
}
\arguments{
\item{userdata}{User's data.}

\item{predictor}{The data of differential equation model}

\item{model}{The model of standardization}

\item{chooseModel}{c(2,1) or c(2,1)}
}
\value{
The plot of all data.
}
\description{
Plot all data
}
