% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver_MultiBiFirst_func.R
\name{Solver_MultiBiFirst_func}
\alias{Solver_MultiBiFirst_func}
\title{Title}
\usage{
Solver_MultiBiFirst_func(data, model, guess, method, guess2, method2)
}
\arguments{
\item{data}{User's data}

\item{model}{Model's class is dataframe.}

\item{guess}{Guess values that contain coefficient and initial values.}

\item{method}{"Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent"}

\item{guess2}{Guess values of multilevel that contain coefficient and initial values.}

\item{method2}{"Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent"}
}
\value{
The result of optimization,SE,RMSE,r-squared,users's data,predictor data and output table.
}
\description{
Title
}
